Project name changes; RenderTextures
The OcularDX11 project was renamed to OcularD3D11. The files belonging
to this project also were renamed from the #####DX11 pattern to
D3D11#####.

This was done because this is really a Direct3D wrapper and not a
DirectX one.

Also added the shell of the RenderTexture (and D3D11RenderTexture)
classes and fixed a bunch of 'unsigned' to 'uint32_t' instances that I
missed.